requirements.txt
paramiko>=3.4.0
typer>=0.12.3
pyyaml>=6.0.1
rich>=13.7.1

